api/requirements.txt
fastapi==0.78.0
uvicorn==0.18.3
pydantic==1.9.2
pydantic-settings==1.2.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==0.20.0
openai==0.27.0 